aspire/requirements.txt
argcomplete==1.12.3
asgiref==3.4.1
aws-psycopg2==1.2.1
black==21.9b0
boto3==1.18.41
botocore==1.21.41
certifi==2021.5.30
cffi==1.14.6
cfn-flip==1.2.3
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.8
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-cors-headers==3.8.0
django-rest-auth==0.9.5
django-s3-storage==0.13.4
djangorestframework==3.12.4
durationpy==0.5
future==0.18.2
hjson==3.0.2
idna==3.2
jmespath==0.10.0
kappa==0.6.0
Markdown==3.3.4
mypy-extensions==0.4.3
oauthlib==3.1.1
pathspec==0.9.0
pep517==0.11.0
pip-tools==6.2.0
placebo==0.9.0
platformdirs==2.3.0
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.2
python-dotenv==0.19.0
python-slugify==5.0.2
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.1
tqdm==4.62.2
troposphere==2.7.0
typing-extensions==3.10.0.2
urllib3==1.26.6
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
zappa==0.53.0
